fix number type MonthlyChargesTranscribing, change int to float32

+++ infrastructure/bunny_request_rest/create_video_library_response.go
@@ -84,6 +84,6 @@ type CreateVideoLibraryResponse struct {
 	TranscribingPriceOverride               *int             `json:"TranscribingPriceOverride"`
 	PremiumEncodingPriceOverride            *int             `json:"PremiumEncodingPriceOverride"`
 	MonthlyChargesTranscribing              float32          `json:"MonthlyChargesTranscribing"`
-	MonthlyChargesPremiumEncoding           int              `json:"MonthlyChargesPremiumEncoding"`
+	MonthlyChargesPremiumEncoding           float32          `json:"MonthlyChargesPremiumEncoding"`
 	MonthlyChargesEnterpriseDrm             int              `json:"MonthlyChargesEnterpriseDrm"`
 }
